+++ src/lib/custom-logger.service.ts
@@ -1,13 +1,13 @@
 import { Inject, Injectable } from '@angular/core';
 import { INGXLoggerConfig } from './config/iconfig';
 import { NGXLogger } from './logger.service';
-import { INGXLoggerConfigEngine, TOKEN_LOGGER_CONFIG_ENGINE } from './config/iconfig-engine';
 import { TOKEN_LOGGER_MAPPER_SERVICE, INGXLoggerMapperService } from './mapper/imapper.service';
 import { TOKEN_LOGGER_METADATA_SERVICE, INGXLoggerMetadataService } from './metadata/imetadata.service';
 import { TOKEN_LOGGER_RULES_SERVICE, INGXLoggerRulesService } from './rules/irules.service';
 import { TOKEN_LOGGER_SERVER_SERVICE, INGXLoggerServerService } from './server/iserver.service';
 import { TOKEN_LOGGER_WRITER_SERVICE, INGXLoggerWriterService } from './writer/iwriter.service';
 import { INGXLoggerMonitor } from './monitor/ilogger-monitor';
+import { INGXLoggerConfigEngineFactory, TOKEN_LOGGER_CONFIG_ENGINE_FACTORY } from './config/iconfig-engine-factory';
 
 // Keeping this to avoid any breaking change for now, this class should be removed later
 
@@ -21,7 +21,7 @@ import { INGXLoggerMonitor } from './monitor/ilogger-monitor';
 export class CustomNGXLoggerService {
 
   constructor(
-    @Inject(TOKEN_LOGGER_CONFIG_ENGINE) private configEngine: INGXLoggerConfigEngine,
+    @Inject(TOKEN_LOGGER_CONFIG_ENGINE_FACTORY) private configEngineFactory: INGXLoggerConfigEngineFactory,
     @Inject(TOKEN_LOGGER_METADATA_SERVICE) private metadataService: INGXLoggerMetadataService,
     @Inject(TOKEN_LOGGER_RULES_SERVICE) private ruleService: INGXLoggerRulesService,
     @Inject(TOKEN_LOGGER_MAPPER_SERVICE) private mapperService: INGXLoggerMapperService,
@@ -45,10 +45,7 @@ export class CustomNGXLoggerService {
     logMonitor?: INGXLoggerMonitor,
     mapperService?: INGXLoggerMapperService,
   ): NGXLogger {
-
-    this.configEngine.updateConfig(config);
-
-    const logger = new NGXLogger(this.configEngine, this.metadataService, this.ruleService, mapperService || this.mapperService, this.writerService, serverService || this.serverService);
+    const logger = new NGXLogger(config, this.configEngineFactory, this.metadataService, this.ruleService, mapperService || this.mapperService, this.writerService, serverService || this.serverService);
 
     if (logMonitor) {
       logger.registerMonitor(logMonitor);