+++ src/lib/logger.service.spec.ts
@@ -1,13 +1,12 @@
-import { DatePipe } from '@angular/common';
 import { inject, TestBed } from '@angular/core/testing';
-import { NGXLoggerConfigEngineMock } from 'testing/src/lib/config-engine.mock';
-import { NGXLoggerMapperServiceMock } from 'testing/src/lib/mapper.service.mock';
-import { NGXLoggerMetadataServiceMock } from 'testing/src/lib/metadata.service.mock';
-import { NGXLoggerRulesServiceMock } from 'testing/src/lib/rules.service.mock';
-import { NGXLoggerServerServiceMock } from 'testing/src/lib/server.service.mock';
-import { NGXLoggerWriterServiceMock } from 'testing/src/lib/writer.service.mock';
+import { NGXLoggerConfigEngineFactoryMock } from 'src/lib/testing/config-engine-factory.mock';
+import { NGXLoggerMapperServiceMock } from 'src/lib/testing/mapper.service.mock';
+import { NGXLoggerMetadataServiceMock } from 'src/lib/testing/metadata.service.mock';
+import { NGXLoggerRulesServiceMock } from 'src/lib/testing/rules.service.mock';
+import { NGXLoggerServerServiceMock } from 'src/lib/testing/server.service.mock';
+import { NGXLoggerWriterServiceMock } from 'src/lib/testing/writer.service.mock';
 import { TOKEN_LOGGER_CONFIG } from './config/iconfig';
-import { TOKEN_LOGGER_CONFIG_ENGINE } from './config/iconfig-engine';
+import { TOKEN_LOGGER_CONFIG_ENGINE_FACTORY } from './config/iconfig-engine-factory';
 import { NGXLogger } from './logger.service';
 import { TOKEN_LOGGER_MAPPER_SERVICE } from './mapper/imapper.service';
 import { TOKEN_LOGGER_METADATA_SERVICE } from './metadata/imetadata.service';
@@ -22,7 +21,7 @@ describe('NGXLogger', () => {
       providers: [
         NGXLogger,
         { provide: TOKEN_LOGGER_CONFIG, useValue: { level: NgxLoggerLevel.ERROR } },
-        { provide: TOKEN_LOGGER_CONFIG_ENGINE, useClass: NGXLoggerConfigEngineMock },
+        { provide: TOKEN_LOGGER_CONFIG_ENGINE_FACTORY, useClass: NGXLoggerConfigEngineFactoryMock },
         { provide: TOKEN_LOGGER_METADATA_SERVICE, useClass: NGXLoggerMetadataServiceMock },
         { provide: TOKEN_LOGGER_RULES_SERVICE, useClass: NGXLoggerRulesServiceMock },
         { provide: TOKEN_LOGGER_MAPPER_SERVICE, useClass: NGXLoggerMapperServiceMock },