+++ src/lib/testing/config-engine.mock.ts
@@ -1,5 +1,7 @@
-import { INGXLoggerConfig, NgxLoggerLevel, INGXLoggerConfigEngine } from 'ngx-logger';
 import { Injectable } from '@angular/core';
+import { INGXLoggerConfig } from 'src/lib/config/iconfig';
+import { INGXLoggerConfigEngine } from 'src/lib/config/iconfig-engine';
+import { NgxLoggerLevel } from 'src/lib/types/logger-level.enum';
 
 @Injectable()
 export class NGXLoggerConfigEngineMock implements INGXLoggerConfigEngine {