+++ src/lib/testing/logger.service.mock.ts
@@ -1,6 +1,8 @@
 import { HttpHeaders, HttpParams } from '@angular/common/http';
-import { INGXLoggerMonitor, INGXLoggerConfig, NgxLoggerLevel } from 'ngx-logger';
 import { Injectable } from "@angular/core";
+import { INGXLoggerConfig } from 'src/lib/config/iconfig';
+import { INGXLoggerMonitor } from 'src/lib/monitor/ilogger-monitor';
+import { NgxLoggerLevel } from 'src/lib/types/logger-level.enum';
 
 // todo bmtheo, there should be an interface or something to make sure this mock sticks to the real API
 @Injectable()