+++ src/lib/testing/metadata.service.mock.ts
@@ -1,5 +1,8 @@
 import { Injectable } from '@angular/core';
-import { INGXLoggerConfig, INGXLoggerMetadata, INGXLoggerMetadataService, NgxLoggerLevel } from 'ngx-logger';
+import { INGXLoggerConfig } from 'src/lib/config/iconfig';
+import { INGXLoggerMetadata } from 'src/lib/metadata/imetadata';
+import { INGXLoggerMetadataService } from 'src/lib/metadata/imetadata.service';
+import { NgxLoggerLevel } from 'src/lib/types/logger-level.enum';
 
 @Injectable()
 export class NGXLoggerMetadataServiceMock implements INGXLoggerMetadataService {